+++ apps/prairielearn/src/ee/auth/lti13/lti13Auth.html.ts
@@ -40,6 +40,7 @@ export const Lti13Test = ({
             <li><b>UID:</b> ${userInfo.uid} (<code>${lti13_instance.uid_attribute}</code>)
             <li><b>UIN:</b> ${userInfo.uin} (<code>${lti13_instance.uin_attribute}</code>)
             <li><b>Name:</b> ${userInfo.name} (<code>${lti13_instance.name_attribute}</code>)
+            <li><b>Email:</b> ${userInfo.email} (<code>${lti13_instance.email_attribute}</code>)
           </ul>
           <h2>All LTI 1.3 claims</h1>
           <pre><code>${JSON.stringify(lti13_claims, null, 2)}</code></pre>