+++ apps/prairielearn/src/ee/auth/saml/router.ts
@@ -1,4 +1,5 @@
-import ERR from 'async-stacktrace';
+import util from 'node:util';
+
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
 import passport from 'passport';
@@ -67,11 +68,13 @@ router.post(
     const uidAttribute = institutionSamlProvider.uid_attribute;
     const uinAttribute = institutionSamlProvider.uin_attribute;
     const nameAttribute = institutionSamlProvider.name_attribute;
+    const emailAttribute = institutionSamlProvider.email_attribute;
 
     // Read the appropriate attributes.
-    const authnUid = uidAttribute ? user.attributes[uidAttribute] : null;
-    const authnUin = uinAttribute ? user.attributes[uinAttribute] : null;
-    const authnName = nameAttribute ? user.attributes[nameAttribute] : null;
+    const authnUid = uidAttribute ? user.attributes[uidAttribute]?.trim() : null;
+    const authnUin = uinAttribute ? user.attributes[uinAttribute]?.trim() : null;
+    const authnName = nameAttribute ? user.attributes[nameAttribute]?.trim() : null;
+    const authnEmail = emailAttribute ? user.attributes[emailAttribute]?.trim() : null;
 
     if (req.body.RelayState === 'test') {
       res.send(
@@ -79,9 +82,11 @@ router.post(
           uid: authnUid,
           uin: authnUin,
           name: authnName,
+          email: authnEmail,
           uidAttribute,
           uinAttribute,
           nameAttribute,
+          emailAttribute,
           attributes: user.attributes,
           resLocals: res.locals,
         }),
@@ -90,6 +95,12 @@ router.post(
     }
 
     // Only perform validation if we aren't rendering the above test page.
+    //
+    // Support for pulling in email from an attribute was added after all initial
+    // attributes, so we can't yet require it to be present. In the future, once
+    // we've specified such an attribute for all institutions, we can assert that
+    // the email attribute mapping and the corresponding value are both present.
+    //
     if (!uidAttribute || !uinAttribute || !nameAttribute) {
       throw new Error('Missing one or more SAML attribute mappings');
     }
@@ -101,6 +112,7 @@ router.post(
       uid: authnUid,
       name: authnName,
       uin: authnUin,
+      email: authnEmail,
       provider: 'SAML',
       institution_id: institutionId,
     };
@@ -114,22 +126,19 @@ router.post(
 
 router.get(
   '/metadata',
-  asyncHandler(async (req, res, next) => {
+  asyncHandler(async (req, res) => {
     const samlProvider = await getInstitutionSamlProvider(req.params.institution_id);
     if (!samlProvider) {
       throw new error.HttpStatusError(404, 'Institution does not support SAML authentication');
     }
 
-    strategy.generateServiceProviderMetadata(
+    const metadata = await util.promisify(strategy.generateServiceProviderMetadata)(
       req,
       samlProvider.public_key,
       samlProvider.public_key,
-      (err, metadata) => {
-        if (ERR(err, next)) return;
-        res.type('application/xml');
-        res.send(metadata);
-      },
     );
+    res.type('application/xml');
+    res.send(metadata);
   }),
 );