+++ apps/prairielearn/src/migrations/20240611160107_assessment_access_rules__number__not_null.sql
@@ -1,0 +1,3 @@
+ALTER TABLE assessment_access_rules
+ALTER COLUMN number
+SET NOT NULL;