+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prairielearn/postgres-tools
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [1de33d9]
+  - @prairielearn/postgres@2.1.0
+
+## 2.0.3
+
+### Patch Changes
+
+- 0f7c90f: Upgrade all JavaScript dependencies
+- Updated dependencies [0f7c90f]
+  - @prairielearn/postgres@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes